+++ packages/orchestration/src/index.ts
@@ -42,7 +42,7 @@ export type {
 
 export { OrchestrationClient } from './orchestration-client.js';
 
-export { buildAzureContentFilter } from './orchestration-filter-utility.js';
+export { buildAzureContentFilter } from './orchestration-util.js';
 
 export { OrchestrationResponse } from './orchestration-response.js';
