+++ packages/orchestration/src/internal.ts
@@ -1,4 +1,4 @@
 export * from './orchestration-client.js';
-export * from './orchestration-filter-utility.js';
+export * from './orchestration-util.js';
 export * from './orchestration-types.js';
 export * from './orchestration-response.js';
