+++ packages/orchestration/src/orchestration-client.test.ts
@@ -9,7 +9,7 @@ import {
   constructCompletionPostRequest,
   OrchestrationClient
 } from './orchestration-client.js';
-import { buildAzureContentFilter } from './orchestration-filter-utility.js';
+import { buildAzureContentFilter } from './orchestration-util.js';
 import { OrchestrationResponse } from './orchestration-response.js';
 import type { CompletionPostResponse } from './client/api/schema/index.js';
 import type {
