+++ packages/orchestration/src/orchestration-util.test.ts
@@ -1,9 +1,9 @@
 import { constructCompletionPostRequest } from './orchestration-client.js';
-import { buildAzureContentFilter } from './orchestration-filter-utility.js';
+import { buildAzureContentFilter } from './orchestration-util.js';
 import type {
   CompletionPostRequest,
   FilteringModuleConfig
-} from './client/api/schema';
+} from './client/api/schema/index.js';
 import type { OrchestrationModuleConfig } from './orchestration-types.js';
 
 describe('filter utility', () => {
